region-contains-region-p: add ellipse vs rect and vs polygon
For simplified case we depend on the fact, that ellipses are convex objects.

For a case when the ellipse is limited with an angle we depend on
region-intersection called on line vs ellipse. If all edges fit, then whole
object fits in the ellipse and none is shattered by the pie edge.

We use same method for polygons.